+++ src/Compute/Compute/ChangeLog.md
@@ -19,8 +19,9 @@
         - Additional information about change #1
 
 -->
-## Version 10.0.0
+## Upcoming Release
 * Deprecated `Get-AzVMSize` 'List Virtual Machine Size' parameter set. 
+* Added new parameters `EnableAutomaticZoneRebalance`, `AutomaticZoneRebalanceStrategy` and `AutomaticZoneRebalanceBehavior` to `New-AzVmssConfig` and `Update-AzVmss` cmdlets for VMSS Automatic Zone Rebalancing. 
 
 ## Version 9.3.0
 * Added new cmdlets `Add-AzVmssSkuProfileVMSize` and `Remove-AzVmssSkuProfileVMSize` to add and remove VM sizes to and from the VMSS SkuProfile.